+++ plat/xilinx/zynqmp/pm_service/pm_api_sys.h
@@ -149,9 +149,9 @@ enum pm_ret_status pm_clock_setdivider(unsigned int clock_id,
 enum pm_ret_status pm_clock_getdivider(unsigned int clock_id,
 				       unsigned int *divider);
 enum pm_ret_status pm_clock_setrate(unsigned int clock_id,
-				    unsigned int rate);
+				    uint64_t rate);
 enum pm_ret_status pm_clock_getrate(unsigned int clock_id,
-				    unsigned int *rate);
+				    uint64_t *rate);
 enum pm_ret_status pm_clock_setparent(unsigned int clock_id,
 				      unsigned int parent_id);
 enum pm_ret_status pm_clock_getparent(unsigned int clock_id,
